+++ pkg/admin/invite_user.go
@@ -1,4 +1,3 @@
-// Package admin provides administrative operations for user and API key management.
 package admin
 
 import (
@@ -6,54 +5,13 @@ import (
 	"crypto/rand"
 	"encoding/hex"
 	"fmt"
-	"time"
-
-	"httpcache/pkg/dbsqlc"
-
-	"github.com/jackc/pgx/v5"
 )
-
-// User represents a user in the system
-type User struct {
-	ID        int64     `json:"id"`
-	Email     string    `json:"email"`
-	CreatedAt time.Time `json:"created_at"`
-}
-
-// APIKey represents an API key assigned to a user
-type APIKey struct {
-	ID        int64     `json:"id"`
-	KeyString string    `json:"key_string"`
-	Status    string    `json:"status"`
-	CreatedAt time.Time `json:"created_at"`
-}
-
-// ServiceQuota represents quota allocation for a service
-type ServiceQuota struct {
-	ServiceName    string `json:"service_name"`
-	InitialQuota   int32  `json:"initial_quota"`
-	RemainingQuota int32  `json:"remaining_quota"`
-}
 
 // InviteNewUserResult represents the result of inviting a new user
 type InviteNewUserResult struct {
 	User          *User           `json:"user"`
 	APIKey        *APIKey         `json:"api_key"`
 	InitialQuotas []*ServiceQuota `json:"initial_quotas"`
-}
-
-// AdminService provides administrative operations
-type AdminService struct {
-	db      *pgx.Conn
-	queries *dbsqlc.Queries
-}
-
-// NewAdminService creates a new admin service
-func NewAdminService(db *pgx.Conn) *AdminService {
-	return &AdminService{
-		db:      db,
-		queries: dbsqlc.New(db),
-	}
 }
 
 // ServiceKeyPrefix is the prefix for service keys
@@ -69,117 +27,9 @@ func generateAPIKey() (string, error) {
 	return ServiceKeyPrefix + hex.EncodeToString(bytes), nil
 }
 
-// CreateNewUser creates a new user in the system.
-// This function will return an error if the user already exists.
-func (as *AdminService) CreateNewUser(ctx context.Context, email string, isServiceKey bool) (int64, error) {
-	// Start transaction
-	tx, err := as.db.Begin(ctx)
-	if err != nil {
-		return 0, fmt.Errorf("failed to begin transaction: %w", err)
-	}
-	defer func() {
-		if rollbackErr := tx.Rollback(ctx); rollbackErr != nil {
-			// Rollback errors are typically expected after successful commits
-			_ = rollbackErr // Acknowledge but don't propagate rollback errors
-		}
-	}()
-
-	// Create queries with transaction context
-	qtx := as.queries.WithTx(tx)
-
-	// Check if user already exists
-	_, err = qtx.GetUserByEmail(ctx, email)
-	if err == nil {
-		return 0, fmt.Errorf("user with email %s already exists", email)
-	}
-
-	// Create new user
-	user, err := qtx.CreateUser(ctx, email)
-	if err != nil {
-		return 0, fmt.Errorf("failed to create user: %w", err)
-	}
-
-	// Commit transaction
-	if err := tx.Commit(ctx); err != nil {
-		return 0, fmt.Errorf("failed to commit transaction: %w", err)
-	}
-
-	return user.ID, nil
-}
-
 // isServiceKey checks if an API key string is a service key based on its prefix
 func isServiceKey(apiKey string) bool {
 	return len(apiKey) > len(ServiceKeyPrefix) && apiKey[:len(ServiceKeyPrefix)] == ServiceKeyPrefix
-}
-
-// AddKeyToUser adds an API key to an existing user and sets up quotas if it's not a service key.
-func (as *AdminService) AddKeyToUser(ctx context.Context, userID int64, apiKey string) error {
-	// Start transaction
-	tx, err := as.db.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
-	defer func() {
-		if rollbackErr := tx.Rollback(ctx); rollbackErr != nil {
-			// Rollback errors are typically expected after successful commits
-			_ = rollbackErr // Acknowledge but don't propagate rollback errors
-		}
-	}()
-
-	// Create queries with transaction context
-	qtx := as.queries.WithTx(tx)
-
-	// Create user API key with quota in "unassigned" status
-	apiKeyRecord, err := qtx.CreateUserAPIKey(ctx, &dbsqlc.CreateUserAPIKeyParams{
-		UserID:    userID,
-		KeyString: apiKey,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to create API key: %w", err)
-	}
-
-	// Update API key status to "assigned"
-	apiKeyRecord, err = qtx.UpdateAPIKeyStatus(ctx, &dbsqlc.UpdateAPIKeyStatusParams{
-		ID:     apiKeyRecord.ID,
-		Status: "assigned",
-	})
-	if err != nil {
-		return fmt.Errorf("failed to update API key status: %w", err)
-	}
-
-	// Initialize quotas only for normal user keys (not service keys)
-	if !isServiceKey(apiKey) {
-		// Get all services to set up quotas
-		services, err := qtx.GetAllServices(ctx)
-		if err != nil {
-			return fmt.Errorf("failed to get services: %w", err)
-		}
-
-		// Initialize quotas for all services
-		for _, service := range services {
-			// Get service details to access default quota
-			serviceDetails, err := qtx.GetServiceByName(ctx, service.Name)
-			if err != nil {
-				return fmt.Errorf("failed to get service details for %s: %w", service.Name, err)
-			}
-
-			_, err = qtx.InitializeKeyServiceQuota(ctx, &dbsqlc.InitializeKeyServiceQuotaParams{
-				ApiKeyID:  apiKeyRecord.ID,
-				ServiceID: service.ID,
-				Available: serviceDetails.DefaultQuota,
-			})
-			if err != nil {
-				return fmt.Errorf("failed to initialize quota for service %s: %w", service.Name, err)
-			}
-		}
-	}
-
-	// Commit transaction
-	if err := tx.Commit(ctx); err != nil {
-		return fmt.Errorf("failed to commit transaction: %w", err)
-	}
-
-	return nil
 }
 
 // InviteNewUser assigns an API key to a new user and optionally sets up initial quotas for all services.
